pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whale_alert"
version = "0.1.0"
description = "A Telegram bot that listens to Whale Alert messages and stores them in TimescaleDB"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "telethon>=1.28.5",
    "asyncpg>=0.27.0",
    "sqlalchemy>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.11.0",
    "python-dateutil>=2.8.2",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "psycopg2-binary>=2.9.10",
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.23.0",
    "tomli>=2.0.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
whale-alert = "whale_alert.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[project.urls]
Homepage = "https://github.com/yourusername/whale_alert"
Documentation = "https://github.com/yourusername/whale_alert#readme"
Issues = "https://github.com/yourusername/whale_alert/issues"
Source = "https://github.com/yourusername/whale_alert"
